create tenative 'share with me' navlinks
